image-host/requirements.txt
flask==1.1.2
Pillow==8.2.0
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4